requirements.txt
asgiref==3.8.1
Django==5.1.3
django-login-required-middleware==0.9.0
sqlparse==0.5.1
